Version 0.08
The main menu was set by me. The icons were created by a photo editor app, they were then imported in unity and I created them as clickable buttons. I wrote a script for the play button, if the user presses it, the game starts, For the quit button, I need to add instructions in my script for the code to close the game. Additionally, for the option and about button, I need to add dialogue boxes containing information and further functions in the game.